move getTooltipTransform to tooltipUtils
