+++ src/widgets/Ivy.Widgets.ActivityHeatmap/frontend/src/tooltipUtils.ts
@@ -16,3 +16,11 @@ export function formatTooltipValue(day: Activity): string {
   const label = day.count === 1 ? "contribution" : "contributions";
   return `${day.count} ${label}`;
 }
+
+export function getTooltipTransform(tooltipDiv: HTMLDivElement | null, tooltipCoordinates: { x: number; y: number }, gridContainer: HTMLDivElement | null): string {
+  if (!tooltipDiv || !gridContainer) return "translate(0px, 0px)";
+  const gridRect = gridContainer.getBoundingClientRect();
+  const xOffset = tooltipCoordinates.x > gridRect.left + gridRect.width / 2 ? -(tooltipDiv.offsetWidth + 20) : 20;
+  const yOffset = tooltipCoordinates.y > window.innerHeight / 2 ? -(tooltipDiv.offsetHeight + 20) : 20;
+  return `translate(${xOffset}px, ${yOffset}px)`;
+}
